matchStatus/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
pymongo[srv]==3.12.1
djongo==1.3.6
pytest-django==4.5.2